+++ androidApp/src/main/java/com/pathok/user/android/design_system/RegularEditText.kt
@@ -1,4 +1,4 @@
-package com.pathok.user.android.designsystem
+package com.pathok.user.android.design_system
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
